fix: clock smoothness, temp accuracy, panel height, and IMU FIFO->NVS pipeline
Fixes found by eyeballing the real display for the first time, plus one
architectural gap from the previous commit's stack-size fix.

Display:
- display_out_time(): only redraw the HH/MM/SS field(s) that actually
  changed instead of the whole string every second — full-string redraw
  pushed ~4x more pixels over SPI than necessary and was visibly sweeping
  left-to-right on this panel (flushBuffer() only sends the dirty bounding
  box). display_clock_time_reset() forces one full redraw (incl. colons)
  whenever the screen was cleared out from under it (mode switches, first
  entry).
- imu_get_temp(): was truncating to integer twice (once via integer /128,
  once via its int16_t return), throwing away all decimal precision. Now
  returns float with proper float division; propagated through
  ui_clock_set_temp()/display_out_temp() (now %.1fF).
- display_out_temp(): value was already Fahrenheit (imu_get_temp() converts
  C->F), just mislabeled "C" — fixed to "F".
- st7735s_compat.h: defHEIGHT was 180, but the physical panel is 128x160 —
  confirmed on hardware that content near row 160-179 (the step-count badge)
  was drawn off the visible glass entirely. Fixed to 160.
- ui.c: init_ui()/change_ui_mode() now clear_display() before the clock
  face draws, and redraw a persistent "WWD-n" title — previously the
  display_phase() boot splash was never cleared, leaving a stray
  background remnant behind the clock digits.

IMU / NVS:
- interrupt.c: real INT1 setup used GPIO_INT_EDGE_RISING; imu_bringup.c's
  diagnostic used GPIO_INT_EDGE_TO_ACTIVE (the config actually verified
  against the FIFO drain rate in imu_notes.md, since INT1 is active-low).
  GPIO interrupt-edge is one hardware setting per pin, so whichever ran
  last silently won. Fixed the real path to match the verified config.
- main.c: re-enabled button_handler_thread (bumped 1024B->2048B stack,
  same reasoning as the previous commit's sensor/ui_refresh fix — it does
  SPI FIFO reads + nvs_log_record() several frames deep under
  NO_OPTIMIZATIONS). This is the only thread that drains the IMU FIFO into
  NVS; with it disabled, NVS_LOG_IMU_SAMPLES logging was silently a no-op
  despite being enabled.

Verified stable on SN2 via dmesg (no USB reboot-loop) and live serial
after each change.

+++ src/hardware/ic/imu/imu.h
@@ -98,7 +98,7 @@ void imu_reg_poll();
 /**
  * @brief function for just getting temperature data from IMU
  */
-int16_t imu_get_temp();
+float imu_get_temp();
 
 
 /**
